+++ Source/Flock/Private/FlockSubsystem.cpp
@@ -12,7 +12,7 @@
 #include "Engine/World.h"
 
 const FString UFlockSubsystem::ApiVersion = TEXT("v1");
-const FString UFlockSubsystem::SdkVersion = TEXT("1.2.0");
+const FString UFlockSubsystem::SdkVersion = TEXT("1.3.0");
 
 UFlockSubsystem* UFlockSubsystem::Get(const UObject* WorldContextObject)
 {